Make the logic of NetClientImpl reusable in NetClientBase
