+++ src/main/java/io/vertx/core/metrics/impl/DummyVertxMetrics.java
@@ -33,7 +33,6 @@ import io.vertx.core.http.HttpServerResponse;
 import io.vertx.core.http.ServerWebSocket;
 import io.vertx.core.http.WebSocket;
 import io.vertx.core.spi.metrics.*;
-import io.vertx.core.net.NetClient;
 import io.vertx.core.net.NetClientOptions;
 import io.vertx.core.net.NetServerOptions;
 import io.vertx.core.net.SocketAddress;
@@ -82,7 +81,7 @@ public class DummyVertxMetrics implements VertxMetrics {
   }
 
   @Override
-  public TCPMetrics createMetrics(NetClient client, NetClientOptions options) {
+  public TCPMetrics createMetrics(NetClientOptions options) {
     return DummyTCPMetrics.INSTANCE;
   }
