+++ src/test/java/io/vertx/test/fakemetrics/FakeVertxMetrics.java
@@ -25,7 +25,6 @@ import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpClientOptions;
 import io.vertx.core.http.HttpServer;
 import io.vertx.core.http.HttpServerOptions;
-import io.vertx.core.net.NetClient;
 import io.vertx.core.net.NetClientOptions;
 import io.vertx.core.net.NetServerOptions;
 import io.vertx.core.net.SocketAddress;
@@ -101,7 +100,7 @@ public class FakeVertxMetrics extends FakeMetricsBase implements VertxMetrics {
     };
   }
 
-  public TCPMetrics<?> createMetrics(NetClient client, NetClientOptions options) {
+  public TCPMetrics<?> createMetrics(NetClientOptions options) {
     return new TCPMetrics<Object>() {
 
       public Object connected(SocketAddress remoteAddress, String remoteName) {
